Split PlanAnalyzer.cs into partial classes (#329)
* Add "Powered by Performance Studio" line on landing page


* Add Darling Data favicon to web app


* Add Open Graph and Twitter Card meta tags for social sharing

Uses the Darling Data barbell logo as hero image when shared on
social media. Also adds meta description for SEO.


* Clarify OG description: in-browser, nothing to install


* Fix Rule 3 severity: CouldNotGenerateValidParallelPlan is actionable

This reason means something in the query blocks parallelism (scalar UDFs,
table variable inserts, etc.) — that's worth a Warning, not Info.


* Expand Rule 3 to cover all NonParallelPlanReason values

Adds human-readable messages for all 25 known reasons. Severity:
- Warning: actionable reasons (UDFs, cursors, table variables, remote
  queries, trace flags, hints, DML OUTPUT, writeback variables)
- Info: passive/environmental (cost below threshold, edition limits,
  memory-optimized tables, upgrade mode, index build edge cases)


* Split PlanAnalyzer.cs into partial classes

Move-only refactor; no behavior changes. PlanAnalyzer.cs (2,238 lines)
split into 5 partials:

  Statement   (439) - AnalyzeStatement (statement-level rules)
  Node        (777) - AnalyzeNodeTree + AnalyzeNode (operator rules)
  Detection   (248) - predicate/pattern helpers (HasNotIn, IsAntiSemi,
                      IsRowstoreScan, IsProbeOnly, DetectNonSargable,
                      IsOrExpansionChain, HasAdaptiveJoin, ...)
  Timing      (345) - operator CPU/elapsed helpers + ScanImpact + memory
  Helpers     (275) - severity overrides, wait labels, format, truncate

Main file now 115 lines — class declaration, regex constants, static
dictionaries (RuleWarningTypes/WarningTypeToRule), static ctor, the
public Analyze() entry point, and the ScanImpact record.

PlanViewer.Web.csproj updated to link the new partial files.

Build clean: 0 errors, 0 warnings on PlanViewer.sln.


---------
